Size optimize the VUnit images
* Delete pip cache saves 15 MB

* Not upgrading pip saves 8 MB

Pip is already installed in /usr/lib/python3/dist-packages, currently version 18.1.
The pip install command will not upgrade that package, but instead install a new one in /usr/local/lib, currently version 20.2.4.
We hardly need a bleeding-edge pip version, and it adds size to the image.

* Do not install ca-certificates

Does not impact size in a noteworthy way.
On the other hand it does not seem like it is necessary for the image.
So in order to keep the dockerfile clean, and only shipping what is necessary, it is removed.